More sophisticated clip detection. No more copying to separate buffer in translate function, as well as delay feature for certain filters. Using all doubles instead of floats.

+++ dsd2pcm.hpp
@@ -60,7 +60,7 @@ public:
     void translate(size_t blockSize,
         const unsigned char *dsdData, ptrdiff_t dsdStride,
         bool lsbitfirst,
-        float *floatData, ptrdiff_t floatStride)
+        double *floatData, ptrdiff_t floatStride)
     {
         dsd2pcm_translate_8to1(handle, blockSize, dsdData, dsdStride,
             lsbitfirst, floatData, floatStride);
